This version might be worse
Both this commit and the previous run the iterative correction without crashing, but the results are a bit off. It works flawlessly in testing with a specific pixel of the sky, but on other regions it does not work. Only on 1 quadrant of the sky it works perfectly. This means there is some angle based transformation that is failing. Need to revise fully the correction.